--- app/controllers/v1/users_controller.rb
@@ -1,71 +1,0 @@
-module V1
-  class UsersController < ApplicationController
-    before_action :set_user, only: [:show, :edit, :update, :destroy]
-
-    # controller for access by api
-    # GET /users/1
-    # GET /users/1.json
-    def show
-      respond_to do |format|
-        if @user.exist?
-          # format.html { redirect_to @user, notice: 'User was successfully created.' }
-          format.json { render :show, status: :created, location: @user }
-        else
-          # format.html { render :new }
-          format.json { render json: @user.errors, status: :unprocessable_entity }
-        end
-      end
-
-    # POST /users
-    # POST /users.json
-    def create
-      @user = User.new(user_params)
-
-      respond_to do |format|
-        if @user.save
-          # format.html { redirect_to @user, notice: 'User was successfully created.' }
-          format.json { render :show, status: :created, location: @user }
-        else
-          # format.html { render :new }
-          format.json { render json: @user.errors, status: :unprocessable_entity }
-        end
-      end
-    end
-
-    # PATCH/PUT /users/1
-    # PATCH/PUT /users/1.json
-    def update
-      respond_to do |format|
-        if @user.update(user_params)
-          # format.html { redirect_to @user, notice: 'User was successfully updated.' }
-          format.json { render :show, status: :ok, location: @user }
-        else
-          format.html { render :edit }
-          format.json { render json: @user.errors, status: :unprocessable_entity }
-        end
-      end
-    end
-
-    # DELETE /users/1
-    # DELETE /users/1.json
-    def destroy
-      @user.destroy
-      respond_to do |format|
-        # format.html { redirect_to users_url, notice: 'User was successfully destroyed.' }
-        format.json { head :no_content }
-      end
-    end
-
-    private
-      # Use callbacks to share common setup or constraints between actions.
-      # get user by apikey
-      def set_user
-        @user = User.find_by(apikey: params[:key])
-      end
-
-      # Never trust parameters from the scary internet, only allow the white list through.
-      def user_params
-        params.fetch(:user, {})
-      end
-  end
-end
